lab2/pyproject.toml
[tool.poetry]
name = "search-labs"
version = "0.1.0"
description = ""
authors = ["Oleksii Dolhov"]

[tool.poetry.dependencies]
python = "^3.10"
certifi = "2022.12.7"
click = "8.1.3"
colorama = "0.4.6"
commonmark = "0.9.1"
elastic-transport = "8.4.0"
elasticsearch = "8.6.1"
pygments = "2.14.0"
python-dotenv = "0.21.1"
rich = "12.6.0"
shellingham = "1.5.0.post1"
typer = "0.7.0"
urllib3 = "1.26.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"